skip existing files +  add logger + add doc string

+++ src/nycdb/dataset.py
@@ -27,8 +27,24 @@ def mkdir(file_path):
 
     
 def download_file(url, dest):
+    """ 
+    Downloads a url and saves the result to the destination path
+    
+    It will creates parent directory of the destination path,
+    if they they don't exist.
+    
+
+    If the destination file exists and is not empty, it assumes the file has
+    already been downloaded and will skip downloading the file accordingly.
+    """
     mkdir(dest)
+
+    if Path(dest).exists() and os.stat(dest).st_size > 0:
+        logging.info("{} has already been downloaded, skipping".format(url))
+        return True
+
     try:
+        logging.info("Downloading {url} to {dest}".format(url=url, dest=dest))
         r = requests.get(url, stream=True)
         with open(dest, 'wb') as f:
             for chunk in r.iter_content(chunk_size=(512 * 1024)):
